bug 1171957 - Create base Extractor, Visitor
The Extractor will be used after HTML is parsed into elements
